Rename third_robot_driver to cirkit_unit03_driver
